requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
kombu==5.0.2
prompt-toolkit==3.0.18
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
